+++ test/support/dev_server_router.ex
@@ -28,6 +28,16 @@ defmodule DevServerRouter do
     }
   )
 
+  forward("/buffered_tiffs",
+    to: IIIFImagePlug.V3,
+    init_opts: %{
+      identifier_to_path_callback: &DevServerHelper.identifier_to_path/1,
+      host: &DevServerHelper.get_host/0,
+      port: &DevServerHelper.get_port/0,
+      temp_dir: :buffer
+    }
+  )
+
   forward("/",
     to: IIIFImagePlug.V3,
     init_opts: %{