+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.0
+- Requested TIF image images are now written to a temporary file instead of to memory by default. 
+- Also a new plug option `:temp_dir` was added to define a custom temporary directory path for the generated TIF files. In memory creation is still possible by passing `:buffer` instead.
+
 ## 0.5.0
 - Changed override options `:port`, `:host` and `:scheme` into callbacks to make values configurable at runtime.
 - Fixed "height" key typo in `info.json`